+++ src/test/java/com/account/banking/AccountBuilder.java
@@ -3,6 +3,8 @@
  */
 package com.account.banking;
 
+import java.math.BigDecimal;
+
 import org.springframework.stereotype.Component;
 
 import com.aspire.banking.domain.Account;
@@ -23,7 +25,7 @@ public class AccountBuilder {
 	private AccountTransfer accountTransfer = null;
 	private Account account = null;
 	
-	 public AccountTransfer createAccountTransfer(long fromId, long toId, double amount, String secret){	
+	 public AccountTransfer createAccountTransfer(long fromId, long toId, BigDecimal amount, String secret){	
 		accountTransfer = new AccountTransfer();
 		accountTransfer.setFromAccount(fromId);
 		accountTransfer.setToAccount(toId);
@@ -32,7 +34,7 @@ public class AccountBuilder {
 		return accountTransfer;
 	}
 
-	public Account createAccount(long i, String fullName, String secret, double balance) {
+	public Account createAccount(long i, String fullName, String secret, BigDecimal balance) {
 		account = new Account();
 		account.setFullName(fullName);
 		account.setId(i);
@@ -41,7 +43,7 @@ public class AccountBuilder {
 		return account;
 	}
 	
-	public Account createAccountWithOutId( String fullName, String secret, double balance) {
+	public Account createAccountWithOutId( String fullName, String secret, BigDecimal balance) {
 		account = new Account();
 		account.setFullName(fullName);
 		account.setSecret(secret);
